fix(core): integrate luau loader with main's config-file and deploy work
Three post-rebase touch-ups to make the luau loader play nicely with
the new code that landed on main while this PR was open:

- The c12 resolve callback's defer-on-explicit-configFile guard now
  reads from a closed-over flag instead of c12Options.configFile,
  because c12 normalizes that field to "bedrock.config" by default
  before invoking the resolver. The previous check always tripped,
  which silently disabled luau auto-discovery when paired with PR #188.
- Move the ConfigError type tests out of index.spec-d.ts (which would
  otherwise breach the 300-line cap) into a colocated
  core/config-error.spec-d.ts, matching the convention used elsewhere.
- Extend deploy.spec-d.ts's ConfigError variant assertion to include
  the new luauRuntimeMissing kind so the deploy() error union stays
  exhaustive.
